Installed socket.io
Although WebSocketsAPI is available through Node, we are using socket.io instead. The socker.io npm package provides built-in features like auto-reconnect, rooms, and fallback to long polling and abstracts some of the complexity of working with the base protocol. Opposed to HTTP, Websocket is stateful and event driven which allows for real-time updates and communication.